style: 프로젝트 초기 설정 - 스토리보드 파일 제거 - 폴더 정리
