+++ ProjectManager/ProjectManager/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
