+++ ProjectManager/ProjectManager/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
