+++ ProjectManager/ProjectManager/Resources/Assets.xcassets/Contents.json
